+++ src/server/components/formRunner/__tests__/helpers.test.ts
@@ -1,10 +1,10 @@
 import supertest from "supertest";
 import * as child_process from "child_process";
 import {
-  isFormRunnerReady,
   startFormRunner,
+  isFormRunnerReady,
   parseFormRunnerWebhookObject,
-} from "../form-runner";
+} from "../helpers";
 
 jest.mock("supertest", () =>
   jest.fn().mockReturnValue({
