+++ src/server/components/formRunner/constants.ts
@@ -1,0 +1,2 @@
+export const FORM_RUNNER_BASE_ROUTE = "/application";
+export const FORM_RUNNER_URL = "localhost:3001";
