+++ src/server/middlewares/index.ts
@@ -5,5 +5,4 @@ export { configureBodyParser } from "./body-parser";
 export { configureViews } from "./views";
 export { configureStaticServer } from "./static";
 export { configureErrorHandlers } from "./error-handlers";
-export { configureFormRunnerProxy } from "./form-runner";
 export { configureCookieParser } from "./cookie-parser";
