--- src/server/services/form-runner/index.ts
@@ -1,10 +1,0 @@
-export {
-  FORM_RUNNER_URL,
-  FORM_RUNNER_BASE_ROUTE,
-  startFormRunner,
-  isFormRunnerReady,
-  parseFormRunnerWebhookObject,
-} from "./form-runner";
-
-export { FormRunnerWebhookData, LawyersFormWebhookData } from "./types";
-export { formRunnerPostRequestSchema } from "./schemas";
